requirements.txt
langchain==0.3.7
langchain-community==0.3.5
langchain-postgres==0.0.12
langchain-openai==0.2.5
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.2
pypdf==5.1.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
